+++ client/src/config/conceptSearchTags.ts
@@ -76,12 +76,6 @@ export const relatedAccordions: IHalLinks = {
     tab: 'objects',
     isSemantic: true,
   },
-  worksAboutAgent: {
-    title: 'Works About or Depicting People and Groups',
-    searchTag: 'lux:conceptDepictedAgentFromRelatedWorks',
-    tab: 'works',
-    jsonSearchTerm: 'aboutAgent',
-  },
   concepts: {
     title: 'Related Concepts',
     searchTag: 'lux:conceptRelatedConcepts',